requirements-test.txt
# Testing dependencies

pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0

# Optional: for creating test textures
Pillow>=10.0.0
